# specs/001-volunteer-care-report/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: StrayHub CRM Care Report API
  version: 0.1.0
  description: >-
    志工日常照護回報與動物近期歷程的 CRM 邊界契約。一般收容所操作由
    FastAPI 重新驗證 Session、Membership、角色與 Active Shelter Context；
    PLATFORM_ADMIN 使用平台級 PLATFORM Scope，且所有跨收容所操作都留下 Audit Record。
...
servers:
  - url: http://localhost:8000
    description: 本機開發
...
security:
  - bearerAuth: []
paths:
...
  /v1/animals/{animalId}/timeline:
    get:
      tags: [Timeline]
      operationId: getAnimalTimeline
      parameters:
        - $ref: "#/components/parameters/AnimalId"
        - name: start_date
          in: query
          schema: { type: string, format: date }
        - name: end_date
          in: query
          schema: { type: string, format: date }
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: 每日摘要與逐筆 Timeline
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TimelineResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFoundOrForbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
...
  /v1/assigned-care-reminders:
    get:
      tags: [Assigned Care]
      operationId: listMyAssignedCareReminders
      parameters:
        - { name: cursor, in: query, schema: { type: [string, 'null'] } }
        - { name: page_size, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
      responses:
        '200': { description: 本人被指派事項的最小資料, content: { application/json: { schema: { $ref: '#/components/schemas/AssignedCareList' } } } }
        '404': { $ref: '#/components/responses/NotFoundOrForbidden' }
  /v1/assigned-care-reminders/{occurrenceId}:
    get:
      tags: [Assigned Care]
      operationId: getMyAssignedCareReminder
      parameters: [{ $ref: '#/components/parameters/OccurrenceId' }]
      responses:
        '200': { description: 本人被指派單次事項的最小資料, content: { application/json: { schema: { $ref: '#/components/schemas/AssignedCareItem' } } } }
        '404': { $ref: '#/components/responses/NotFoundOrForbidden' }
  /v1/assigned-care-reminders/{occurrenceId}/actions:
    post:
      tags: [Assigned Care]
      operationId: actOnMyAssignedCareReminder
      parameters:
        - $ref: '#/components/parameters/OccurrenceId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AssignedCareAction' } } } }
      responses:
        '200': { description: 志工完成或略過本人指派事項, content: { application/json: { schema: { $ref: '#/components/schemas/AssignedCareMutation' } } } }
        '404': { $ref: '#/components/responses/NotFoundOrForbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 簽署的短效 Access Token；每次受保護 Request 仍必須通過 Server-side Session 與 Organization Scope 驗證。
  parameters:
...
    AnimalId:
      {
        name: animalId,
        in: path,
        required: true,
        schema: { type: string, format: uuid },
      }
...
    OccurrenceId: { name: occurrenceId, in: path, required: true, schema: { type: string, format: uuid } }
...
    IdempotencyKey:
      {
        name: Idempotency-Key,
        in: header,
        required: true,
        schema: { type: string, minLength: 1 },
      }
...
    Page:
      {
        name: page,
        in: query,
        schema: { type: integer, minimum: 1, default: 1 },
      }
    PageSize:
      {
        name: page_size,
        in: query,
        schema: { type: integer, minimum: 1, maximum: 100, default: 20 },
      }
...
  responses:
...
    Unauthorized:
      description: 未通過身分驗證
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/ErrorResponse" } },
        }
    Forbidden:
      description: 無權限或無法存取，不能洩漏資源存在性
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/ErrorResponse" } },
        }
    NotFoundOrForbidden:
      description: 不存在或未授權的一致結果
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/ErrorResponse" } },
        }
    Conflict:
      description: Idempotency、stale version、operation payload 或資料狀態衝突
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/ErrorResponse" } },
        }
    UnprocessableEntity:
      description: 日期區間或欄位驗證失敗
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/ErrorResponse" } },
        }
...
  schemas:
    ErrorResponse:
      type: object
      required: [code, message, request_id]
      properties:
        code: { type: string }
        message: { type: string }
        request_id: { type: string }
        operation_id: { type: [string, "null"], format: uuid }
        details: { type: object, additionalProperties: true }
...
    TimelineResponse:
      type: object
      required: [days]
      properties:
        days:
          type: array
          items:
            type: object
            required: [date, has_report, report_count]
            properties:
              date: { type: string, format: date }
              has_report: { type: boolean }
              report_count: { type: integer, minimum: 0 }
              no_report_label: { type: [string, "null"], example: 當日無回報 }
              reports:
                {
                  type: array,
                  items: { $ref: "#/components/schemas/TimelineReport" },
                }
    TimelineReport:
      type: object
      required:
        [
          id,
          submitted_at,
          volunteer_user_id,
          animal_name_snapshot,
          status,
          ai_job_status,
          observations,
          observation_snapshots,
          media_ids,
        ]
      properties:
        id: { type: string, format: uuid }
        submitted_at: { type: string, format: date-time }
        volunteer_user_id: { type: string, format: uuid }
        animal_name_snapshot: { type: string }
        shelter_number_snapshot: { type: [string, "null"] }
        note: { type: [string, "null"] }
        observations: { type: object, additionalProperties: { type: string } }
        observation_snapshots:
          type: [object, "null"]
          additionalProperties:
            type: object
            additionalProperties: { type: string }
        status: { type: string, enum: [saved, amended, archived] }
        ai_job_status: { type: string }
        media_ids:
          type: array
          items: { type: string, format: uuid }
...
    ReminderType:
      type: string
      enum: [medication, follow_up, weight, vaccination, examination, other]
...
    AssignedCareAnimal:
      type: object
      additionalProperties: false
      required: [id, name, shelter_number, photo_url]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        shelter_number: { type: [string, 'null'] }
        photo_url: { type: [string, 'null'], format: uri }
    AssignedCareItem:
      type: object
      additionalProperties: false
      required: [occurrence_id, version, status, animal, reminder_type, title, instructions, display_local_at, can_complete, can_skip]
      properties:
        occurrence_id: { type: string, format: uuid }
        version: { type: integer, minimum: 0 }
        status: { type: string, enum: [pending, completed, skipped, cancelled] }
        animal: { $ref: '#/components/schemas/AssignedCareAnimal' }
        reminder_type: { $ref: '#/components/schemas/ReminderType' }
        title: { type: string }
        instructions: { type: string }
        display_local_at: { type: string }
        can_complete: { type: boolean }
        can_skip: { type: boolean }
    AssignedCareList:
      type: object
      additionalProperties: false
      required: [items, next_cursor]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/AssignedCareItem' } }
        next_cursor: { type: [string, 'null'] }
    AssignedCareAction:
      type: object
      additionalProperties: false
      required: [action, expected_version]
      properties:
        action: { type: string, enum: [complete, skip] }
        expected_version: { type: integer, minimum: 0 }
        actual_completed_at: { type: [string, 'null'], format: date-time }
        result_note: { type: [string, 'null'], maxLength: 5000 }
        reason: { type: [string, 'null'], maxLength: 2000 }
    AssignedCareMutation:
      type: object
      additionalProperties: false
      required: [occurrence, action_id, action_type, acted_at, recorded_at, actual_completed_at]
      properties:
        occurrence: { $ref: '#/components/schemas/AssignedCareItem' }
        action_id: { type: string, format: uuid }
        action_type: { type: string, enum: [completed, skipped] }
        acted_at: { type: string, format: date-time }
        recorded_at: { type: [string, 'null'], format: date-time }
        actual_completed_at: { type: [string, 'null'], format: date-time }